+++ js/sidepanel.js
@@ -22,9 +22,21 @@ document.addEventListener('DOMContentLoaded', function () {
 
     /* Set for initial active tab when open the sidepanel */
 
+    const params = new URLSearchParams(window.location.search);
+    let tabId = parseInt(params.get('tabid') || 0, 10);
+    console.log(tabId);
+    if (tabId){
+        chrome.tabs.get(tabId, async tab =>{
+            tabCallback(tab);
+        });
+    } else {
+        chrome.tabs.query({active: true, currentWindow: true}, async tabs =>{
+            tabCallback(tabs[0]);
+        });
+    }
 
-    chrome.tabs.query({active: true, currentWindow: true}, async tabs =>{
-        const tab = tabs[0];
+
+    async function tabCallback(tab){
         tabId = tab.id;
         instance = (new URL(tab.url)).host.replace(".service-now.com", "");
         snuInstanceTagConfig = await getFromSyncStorage("instancetag");
@@ -40,8 +52,8 @@ document.addEventListener('DOMContentLoaded', function () {
             tagTextDoubleclick.value = snuInstanceTagConfig.tagTextDoubleclick;
             tagCommand.value = snuInstanceTagConfig.tagCommand;
             tagCommandShift.value = snuInstanceTagConfig.tagCommandShift;
-        }        
-    });
+        }   
+    }
 
     document.querySelectorAll('input').forEach(inp => { 
         ['change','keyup'].forEach( async evtname => {